spec 009 R3 (T021): physical reorg — 14 stage modules to admesh/_stages/
Executes the package layout side of the Constitution Amendment
(specs/009-release-readiness-for-0.1.0/CONSTITUTION-AMENDMENT.md, proposed
Article VIII). The 14 faithful-port stage modules now live under
admesh/_stages/; their original paths are preserved as backward-compatible
stub re-exports until ADMESH 1.0.0.

Moves (git mv preserves history):
  admesh/background_grid.py  -> admesh/_stages/background_grid.py
  admesh/bathymetry.py       -> admesh/_stages/bathymetry.py
  admesh/boundary.py         -> admesh/_stages/boundary.py
  admesh/curvature.py        -> admesh/_stages/curvature.py
  admesh/distance.py         -> admesh/_stages/distance.py
  admesh/distmesh.py         -> admesh/_stages/distmesh.py
  admesh/domains.py          -> admesh/_stages/domains.py
  admesh/dominate_tide.py    -> admesh/_stages/dominate_tide.py
  admesh/in_polygon.py       -> admesh/_stages/in_polygon.py
  admesh/inpaint.py          -> admesh/_stages/inpaint.py
  admesh/medial_axis.py      -> admesh/_stages/medial_axis.py
  admesh/mesh_size.py        -> admesh/_stages/mesh_size.py
  admesh/quality.py          -> admesh/_stages/quality.py
  admesh/routine.py          -> admesh/_stages/routine.py

New files:
  admesh/_stages/__init__.py — subpackage docstring + empty __all__.
  admesh/<name>.py x14       — backward-compat stub re-exports
                                (covers wildcard public names AND
                                underscore-prefixed helpers used by
                                existing tests).

Internal cross-imports updated to canonical paths (admesh._stages.X) in:
  - admesh/_stages/{bathymetry,boundary,curvature,distmesh,medial_axis,
    mesh_size,routine}.py — sibling stage references.
  - admesh/api.py            — Domain, mesh_quality, routine.triangulate.
  - admesh/__init__.py       — re-exports mesh_quality/right_iso_quality.
  - admesh/valence.py        — mesh_quality.

Stub re-export pattern (per file):
    from admesh._stages.<name> import *
    from admesh._stages import <name> as _src
    globals().update({k: v for k, v in vars(_src).items()
                       if not k.startswith('__')})
    del _src

The second block covers underscore-prefixed names (_signed_area,
_boundary_cleanup, _HAVE_NUMBA, _seed_boundary_1d, ...) that existing
tests reach into for whitebox assertions. 1.0.0 will drop these stubs
and the tests should migrate to admesh._stages.* by then.

Verified:
  - python3 -c "from admesh.curvature import apply_curvature as a;
    from admesh._stages.curvature import apply_curvature as b;
    assert a is b"  passes (single function object, two paths).
  - pytest -m "not slow" -q  passes: 362 passed, 4 deselected, 12 skipped.
  - bash scripts/pre_tag_check.sh  all 9 gates pass.
