+++ admesh/_stages/background_grid.py
@@ -1,0 +1,7 @@
+"""background_grid — port of 02_Create_Background_Grid/CreateBackgroundGrid.m.
+
+MATLAB source: github.com/domattioli/QuADMesh-MATLAB @ 19b2eb9,
+path 01_ADMESH_Library/.
+
+Status: stub. See PROJECT_PLAN.md for the phase that implements this.
+"""
