+++ admesh/api.py
@@ -296,7 +296,7 @@ class Domain:
         #     walk in _derive_boundary_segments which now produces closed rings even for
         #     real-world ADCIRC meshes with pinch-point boundary nodes.
         from scipy.spatial import cKDTree
-        from admesh.in_polygon import in_polygon as _in_polygon
+        from admesh._stages.in_polygon import in_polygon as _in_polygon
 
         outer_ring_nodes = bc_segments[0].node_ids
         outer_ring_pts = mesh.nodes[outer_ring_nodes]
@@ -646,9 +646,9 @@ def triangulate(
         domain = _load_domain_from_source(domain)
     # Lazy imports — keeps `import admesh` cheap and avoids a hard
     # dependency cycle with the faithful-port modules at import time.
-    from admesh.domains import Domain as _PortDomain
-    from admesh.quality import mesh_quality
-    from admesh.routine import triangulate as _routine_triangulate
+    from admesh._stages.domains import Domain as _PortDomain
+    from admesh._stages.quality import mesh_quality
+    from admesh._stages.routine import triangulate as _routine_triangulate
 
     # Resolve h0 from h_max, falling back to a fraction of the bbox diagonal.
     if h_max is None:
